refactor: 컴포넌트명 및 파일명 수정
